Add countdown, coconut types, score popups, pause menu, streak bonus in time mode
- 3-2-1 GO! countdown before game starts; input blocked during countdown
- Three new coconut types: fast (red, x1.4 speed, +20pts), giant (purple, x0.7 speed, +30pts), golden (gold, +50pts)
- Smooth canvas rotation animation for falling coconuts (unique spin rate per coconut)
- Score popups float up from landing point, colour-coded by coconut type
- Pause menu redesigned: RESUME and MAIN MENU buttons with clear tap targets
- Streak bonus enabled in time trial mode (+5..+50 per consecutive catch)
- Classic mode keeps flat points (10/20/30/50 by type, no streak inflation)
- goToMenu() added for pause exit; countdown runnable cleaned up on detach/end
- Base fall speed tuned to 20-27 px/frame (~1.5-2.5 s fall time)

+++ app/src/main/java/kzn/mrcrabs/project/Explosion.java
@@ -8,7 +8,7 @@ public class Explosion {
 
     private static Bitmap[] frames;
     int explosionFrame = 0;
-    int explosionX, explosionY;
+    float explosionX, explosionY;
 
     public Explosion(Context context) {
         if (frames == null) {
